refactor(admin-ui): move version and updates into System Info
Version sits with the other build/runtime facts, so relocate the version row,
check-updates button and update dialog there. General now only holds the
developer mode permission switch.

Also drop the deprecated #body-content slot in favour of the Dialog default slot.
